backend_django/django/espaceclienttopnet/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
platformdirs==4.2.2
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.2
